refactor(cms): move Card render helpers to utils
Keep constants.py for choices only; class and attribute helpers live in taccsite_card/utils.py.
